made client rendering and mobile page
