+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "next": "14.0.1",
     "react": "^18",
+    "react-device-detect": "^2.2.3",
     "react-dom": "^18",
     "react-draggable": "^4.4.6",
     "react-resizable": "^3.0.5"
