+++ src/components/Mobile.js
@@ -1,0 +1,7 @@
+export default function Mobile () {
+    return(
+        <div className="flex flex-col w-[100%] h-[100%] p-1 bg-white overflow-auto">
+            <p>Sorry! This website is only available in mobile</p>
+        </div>
+    )
+}
